refactor(engine): extract daemon runner functions to src/daemon/runner/
All remaining runner orchestration functions extracted from workflow-runner.ts:
- runner/construct-tools.ts: constructTools() (uses injected runWorkflowFn)
- runner/finalize-session.ts: finalizeSession()
- runner/pre-agent-session.ts: buildPreAgentSession()
- runner/agent-loop-runner.ts: buildAgentReadySession(), buildAgentCallbacks(),
  buildUserMessage(), buildTurnEndSubscriber(), TurnEndSubscriberContext,
  runAgentLoop()

workflow-runner.ts: 3,909 lines (original) -> 418 lines (89% reduction).
Now contains only runWorkflow() and re-exports.

Architecture test updated: constructTools no-state-access test now scans
runner/construct-tools.ts (where the function lives).

378/378 test files, 5860 tests pass. Build clean.
